requirements.txt
pyTelegramBotAPI
Wikipedia-API
wikipedia
attr
python-dotenv
